+++ platform/viewer/src/connectedComponents/ConnectedLayoutButton.js
@@ -2,22 +2,24 @@ import { LayoutButton } from '@ohif/ui';
 import OHIF from '@ohif/core';
 import { connect } from 'react-redux';
 
-const { setLayout } = OHIF.redux.actions;
+const { setLayout, setViewportActive } = OHIF.redux.actions;
 
 const mapStateToProps = state => {
   return {
     currentLayout: state.viewports.layout,
+    activeViewportIndex: state.viewports.activeViewportIndex,
   };
 };
 
 const mapDispatchToProps = dispatch => {
   return {
     // TODO: Change if layout switched becomes more complex
-    onChange: (selectedCell, currentLayout) => {
-      let viewports = [];
-      const rows = selectedCell.row + 1;
-      const columns = selectedCell.col + 1;
-      const numViewports = rows * columns;
+    onChange: (selectedCell, currentLayout, activeViewportIndex) => {
+      const viewports = [];
+      const numRows = selectedCell.row + 1;
+      const numColumns = selectedCell.col + 1;
+      const numViewports = numRows * numColumns;
+
       for (let i = 0; i < numViewports; i++) {
         // Hacky way to allow users to exit MPR "mode"
         const viewport = currentLayout.viewports[i];
@@ -27,14 +29,19 @@ const mapDispatchToProps = dispatch => {
         }
 
         viewports.push({
-          height: `${100 / rows}%`,
-          width: `${100 / columns}%`,
           plugin,
         });
       }
       const layout = {
+        numRows,
+        numColumns,
         viewports,
       };
+
+      const maxActiveIndex = numViewports - 1;
+      if (activeViewportIndex > maxActiveIndex) {
+        dispatch(setViewportActive(0));
+      }
 
       dispatch(setLayout(layout));
     },
@@ -43,12 +50,13 @@ const mapDispatchToProps = dispatch => {
 
 const mergeProps = (propsFromState, propsFromDispatch) => {
   const onChangeFromDispatch = propsFromDispatch.onChange;
-  const { currentLayout } = propsFromState;
+  const { currentLayout, activeViewportIndex } = propsFromState;
 
   return {
-    onChange: (selectedCell) => onChangeFromDispatch(selectedCell, currentLayout)
+    onChange: selectedCell =>
+      onChangeFromDispatch(selectedCell, currentLayout, activeViewportIndex),
   };
-}
+};
 
 const ConnectedLayoutButton = connect(
   mapStateToProps,