+++ platform/viewer/src/connectedComponents/ConnectedPluginSwitch.js
@@ -1,9 +1,9 @@
-import OHIF from "@ohif/core";
-import PluginSwitch from "./PluginSwitch.js";
-import { commandsManager } from "./../App.js";
-import { connect } from "react-redux";
+// import OHIF from '@ohif/core';
+import PluginSwitch from './PluginSwitch.js';
+import { commandsManager } from './../App.js';
+import { connect } from 'react-redux';
 
-const { setLayout } = OHIF.redux.actions;
+// const { setLayout } = OHIF.redux.actions;
 
 const mapStateToProps = state => {
   const { activeViewportIndex, layout, viewportSpecificData } = state.viewports;
@@ -11,17 +11,17 @@ const mapStateToProps = state => {
   return {
     activeViewportIndex,
     viewportSpecificData,
-    layout
+    layout,
   };
 };
 
-const mapDispatchToProps = dispatch => {
-  return {
-    setLayout: data => {
-      dispatch(setLayout(data));
-    }
-  };
-};
+// const mapDispatchToProps = dispatch => {
+//   return {
+//     setLayout: data => {
+//       dispatch(setLayout(data));
+//     }
+//   };
+// };
 
 /*function setSingleLayoutData(originalArray, viewportIndex, data) {
   const viewports = originalArray.slice();
@@ -42,17 +42,15 @@ const mergeProps = (propsFromState, propsFromDispatch, ownProps) => {
   const mpr = () => {
         commandsManager.runCommand("mpr2d");
         togglePlugSwitchStatus();
-      }
-  ;
-
+  };
   return {
-    mpr
+    mpr,
   };
 };
 
 const ConnectedPluginSwitch = connect(
   mapStateToProps,
-  mapDispatchToProps,
+  null, // mapDispatchToProps
   mergeProps
 )(PluginSwitch);